requirements.txt
appnope==0.1.2
asgiref==3.3.4
astroid==2.5.3
autopep8==1.5.6
backcall==0.2.0
coloredlogs==15.0
cowsay==4.0
dbs @ file:///Users/kenedinovriansyah7/fires/ideal-daily/ds/dbs/dist/dbs-0.1.tar.gz
decorator==5.0.7
Django==3.2
django-extensions==3.1.3
django-filter==2.4.0
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Faker==8.1.0
humanfriendly==9.1
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
parso==0.8.2
pexpect==4.8.0
phonenumbers==8.12.21
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.8.1
PyJWT==1.7.1
pylint==2.7.4
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
wrapt==1.12.1
